Remove Monitor header from Network
